# medi-care-application/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Management System
  description: |-
    This helps to manage the hospital
  termsOfService: http://swagger.io/terms/
  contact:
    email: aravind@monkeeys.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: localhost:8080/hospital
...
paths:
...
  /v1/api/dashboard:
    get:
      tags:
        - Dashboard
      summary: Get summary details for logined user
      description: Return the list values
      operationId: getSummaryForLoginedUser
      parameters:
        - name: userId
          in: query
          description: User Id
          required: false
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Summary'
        '400':
          description: Invalid user id
  /v1/api/doctor:
    post:
      tags:
        - Doctor
      summary: Creation of new doctor
      description: Return the list values
      operationId: createNewDoctor
      requestBody:
        description: Update an existent pet in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Doctor'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
        '400':
          description: Invalid user id
    get:
      tags:
        - Doctor
      summary: Get Doctors
      description: Get doctors list by params
      operationId: getDoctors
      parameters:
        - name: page
          in: query
          description: Page number
          required: true
          explode: true
          schema:
            type: number
            default: 0
        - name: count
          in: query
          description: Data count
          required: true
          explode: true
          schema:
            type: number
            default: 5
        - name: doctorId
          in: query
          description: Doctor Id
          required: false
          explode: true
          schema:
            type: number
            default: 4598
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Doctor'
        '400':
          description: Invalid user id
  /v1/api/doctor/{id}:
    get:
      tags:
        - Doctor
      summary: Creation of new doctor
      description: Return the list values
      operationId: getDoctorsById
      parameters:
        - name: id
          in: path
          description: User Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
        '400':
          description: Invalid user id
    put:
      tags:
        - Doctor
      summary: Creation of new doctor
      description: Return the list values
      operationId: updateDoctor
      requestBody:
        description: Update an existent pet in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Doctor'
        required: true
      parameters:
        - name: id
          in: path
          description: User Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Doctor'
        '400':
          description: Invalid user id
    delete:
      tags:
        - Doctor
      summary: Delete Operation
      description: Return the list values
      operationId: deleteDoctorsById
      parameters:
        - name: id
          in: path
          description: User Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                example: "Doctor removed successfully"
        '400':
          description: Invalid user id
...
  /v1/api/appointments:
    post:
      tags:
        - Appointment
      summary: Creation of new Appointment
      description: Return the list values
      operationId: bookAppointment
      requestBody:
        description: Appointment creation payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Invalid user id
    get:
      tags:
        - Appointment
      summary: Get Appointments
      description: Get Appointment list by params
      operationId: getAppointments
      parameters:
        - name: page
          in: query
          description: Page number
          required: true
          explode: true
          schema:
            type: number
            default: 0
        - name: count
          in: query
          description: Data count
          required: true
          explode: true
          schema:
            type: number
            default: 5
        - name: appointmentDate
          in: query
          description: ppointment Date in mm-dd-yyyy format
          required: false
          explode: true
          schema:
            type: string
            format: date
            example: '12-31-2024'
        - name: patientId
          in: query
          description: Patient ID
          required: false
          explode: true
          schema:
            type: string
            default: 45689
        - name: doctorId
          in: query
          description: Doctor ID
          required: false
          explode: true
          schema:
            type: string
            default: 45689
        - name: phoneNumber
          in: query
          description: Patient Phone number
          required: false
          explode: true
          schema:
            type: string
            default: 45689
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '400':
          description: Invalid user id
  /v1/api/appointments/{id}:
    get:
      tags:
        - Appointment
      summary: Get appointment by id
      description: Appoinment details
      operationId: getAppoinmentById
      parameters:
        - name: id
          in: path
          description: Appoinment Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Invalid Appointment id
    put:
      tags:
        - Appointment
      summary: Update Appointment
      description: Update Appointment data
      operationId: updateAppointment
      requestBody:
        description: Update an existent user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
        required: true
      parameters:
        - name: id
          in: path
          description: Appointment Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Invalid Appointment id
    delete:
      tags:
        - Appointment
      summary: Delete Appointment
      description: Delete Appointment by id
      operationId: deleteAppointmentById
      parameters:
        - name: id
          in: path
          description: Appointment Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                example: "Appointment removed successfully"
        '400':
          description: Invalid Appointment id
...
  /v1/api/invoices:
    post:
      tags:
        - Invoices
      summary: Creation of new invoices
      description: Return the list values
      operationId: createInvoices
      requestBody:
        description: Payroll creation Invoices
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoices'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoices'
        '400':
          description: Invalid Payroll
    get:
      tags:
        - Invoices
      summary: Get Invoices
      description: Get Invoices list by params
      operationId: getInvoices
      parameters:
        - name: page
          in: query
          description: Page number
          required: true
          explode: true
          schema:
            type: number
            default: 0
        - name: count
          in: query
          description: Data count
          required: true
          explode: true
          schema:
            type: number
            default: 5
        - name: month
          in: query
          description: Month of invoice
          required: false
          explode: true
          schema:
            type: string
            format: date
            example: JAN
        - name: year
          in: query
          description: invoice year
          required: false
          explode: true
          schema:
            type: integer
            default: 45689
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoices'
        '400':
          description: Invalid user id
  /v1/api/invoices/{id}:
    get:
      tags:
        - Invoices
      summary: Get Invoices by id
      description: Appoinment details
      operationId: getInvoicesById
      parameters:
        - name: id
          in: path
          description: Invoices Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRoll'
        '400':
          description: Invalid Invoices id
    put:
      tags:
        - Invoices
      summary: Update Invoices
      description: Update Invoices data
      operationId: updateInvoices
      requestBody:
        description: Update an existent Invoices
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoices'
        required: true
      parameters:
        - name: id
          in: path
          description: PayRoll Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoices'
        '400':
          description: Invalid Invoices id
    delete:
      tags:
        - Invoices
      summary: Delete Invoices
      description: Delete Invoices by id
      operationId: deleteInvoicesById
      parameters:
        - name: id
          in: path
          description: Invoices Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                example: "Invoices removed successfully"
        '400':
          description: Invalid Invoices id
...
components:
  schemas:
    Summary:
      type: object
      properties:
        total_doctors:
          type: integer
          format: int64
          example: 10
        total_patients:
          type: integer
          format: int64
          example: 10
        current_month_cxpense:
          type: number
          example: 15963.50
        pending_payaments:
          type: number
          example: 10
...
    UserBasiInfo:
      type: object
      properties:
        user_id:
          type: string
          example: fb17eecf-5ec6-4c13-a5bd-5ce1451472d0
        first_name:
          type: string
          example: Sastika
        last_name:
          type: string
          example: Aravind
        role:
          $ref: '#/components/schemas/Role'
        gender:
          type: string
          example: Male
          enum:
            - Male
            - FeMale
            - Others
        employment_details:
          $ref: '#/components/schemas/EmploymentDetails'
    Address:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        address_category:
          type: string
          example: Personal
          enum:
            - CURRENT
            - OFFICE
            - PERMANENT
        country:
          type: string
          example: India
        state:
          type: string
          example: Tamilnadu
        city:
          type: string
          example: Coimabtore
        street:
          type: string
          example: 121,Balaji Nagar
        pin:
          type: string
          example: 624601
        geo_location:
          type: string
          example: India
    Contact:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        address:
          $ref: '#/components/schemas/Address'
        phone:
          type: string
          example: 7010192967
        email:
          type: string
          example: aravind@gmail.com
    ProfessionalInformation:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        medical_license_number:
          type: string
          example: 8975469532
          description: Only required for doctors
        years_of_experience:
          type: number
          example: 10.6
        qualifications:
          type: string
          example: MBBS.MS
    Doctor:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        basic_info:
          $ref: '#/components/schemas/UserBasiInfo'
        contact:
          $ref: '#/components/schemas/Contact'
        professional_information:
          $ref: '#/components/schemas/ProfessionalInformation'
    Patients:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        basic_info:
          $ref: '#/components/schemas/UserBasiInfo'
        contact:
          $ref: '#/components/schemas/Contact'
    Appointment:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        appointment_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "Appointment Booking Date"
          example: "2024-11-05T00:00:00Z"
        appointment_time:
          type: string
          example: "10:00 AM"
        patient:
          $ref: '#/components/schemas/Patients'
        doctor:
          $ref: '#/components/schemas/Doctor'
    Role:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        name:
          type: string
          example: ROLE_USER
        description:
          type: string
          example: "User role"
    EmploymentDetails:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        joining_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "User joining date."
          example: "2024-11-05T00:00:00Z"
        relieving_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "User joining date."
          example: "2024-11-05T00:00:00Z"
        department:
          type: string
          example: "CLINIC_STAFF"
        salary:
          type: number
          example: 18500
        payroll:
          $ref: '#/components/schemas/PayRoll'
    PayRoll:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        user:
          $ref: '#/components/schemas/UserBasiInfo'
        payment_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "Appointment Booking Date"
          example: "2024-11-05T00:00:00Z"
        gross_amount:
          type: number
          format: float
        deductions:
          type: number
          format: float
        net_amount:
          type: number
          format: float
...
    Invoices:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        invoice_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "Invoice Date"
          example: "2024-11-05T00:00:00Z"
        total_amount:
          type: number
          format: float
          example: 78965
        deductions:
          type: number
          example: 1596.32
        deduction_summary:
          type: array
          items:
            $ref: '#/components/schemas/DeductionSummary'
        notes:
          type: string
          example: invoice for the employees salary
        reference_id:
          type: number
          example: 158
          description: user id or reference id of the table based on the type.If payroll its mean payroll history id other wise user id(vendor).
        receiver_name:
          type: string
          example: aravind
        invoice_reference_type:
          type: string
          example: PAYROLL
          enum:
            - PAYROLL
            - EXPENSES
            - PURCHASE
            - SALES
            - BILL
            - OTHERS
        status:
          type: string
          example: UN_PAID
          enum:
            - UN_PAID
            - PAID
            - PROCESSING
...
    DeductionSummary:
      type: object
      properties:
        reason:
          type: string
          example: TDS
        amount:
          type: number
          example: 1800.30
...